[TEST] Improve tests
